Mybooks deleted and MyBookBoard modefied
