+++ src/components/MyPage/CSS/MyPageProFile.css
@@ -13,7 +13,7 @@
 
 .ProFileDetail {
     position: relative;
-    left: 35%;
+    left: 18%;
     font-size:200%;
 }
