+++ src/components/MyPage/MyBookBoard.js
@@ -1,5 +1,7 @@
 import React, {Component} from "react"
 import { Link } from 'react-router-dom';
+import server_url from '../../url.json'
+import Image from 'react-image-resizer'
 
 import './CSS/MyBookBoard.css'
 
@@ -9,15 +11,16 @@ class MyBookBoard extends Component {
         console.log(this.props)
         return(
             <Link to={{
-                pathname : `/postdetail/${this.props.url}`,
-                state : {
+                pathname : `/postdetail/${this.props.postID}`,
+               /*  state : {
                     imgUrl : `https://picsum.photos/300/300?image=${this.props.url}`,
                     username : this.props.author,
-                }
+                } */
             }}>
                 <div className ='MyBookBoard'>
                     {console.log('BookBoard component에서 this.props 찍는중', this.props)}
-                    <img className = 'likeThumbnail' src = {`https://picsum.photos/300/300?image=${this.props.url}`} alt='bookcover' />
+                    <Image className = 'likeThumbnail' src = {`http://${server_url}:3000/upload/${this.props.image}`} alt='bookcover' width={200} height={200}/>
+                    <h1>{this.props.title}</h1>
                 </div>
             </Link>
