--- src/components/MyPage/MyBooks.css
@@ -1,3 +1,0 @@
-.myBooks {
-  background-color: blueviolet;
-}
